Added user verification to various controller methods.

+++ lib/stormchat_web/controllers/location_controller.ex
@@ -20,12 +20,24 @@ defmodule StormchatWeb.LocationController do
     end
   end
 
+  # creates a saved location for the verified user
   def create(conn, %{"location" => location_params}) do
-    with {:ok, %Location{} = location} <- Locations.create_location(location_params) do
-      conn
-      |> put_status(:created)
-      |> put_resp_header("location", location_path(conn, :show, location))
-      |> render("show.json", location: location)
+    token = assigns(conn, :token)
+
+    case Phoenix.Token.verify(conn, "auth token", token, max_age: 86400) do
+      {:ok, user_id} ->
+        # make sure a verified user can only create locations for themselves
+        location_params = Map.put(location_params, :user_id, user_id)
+
+        with {:ok, %Location{} = location} <- Locations.create_location(location_params) do
+          conn
+          |> put_status(:created)
+          |> put_resp_header("location", location_path(conn, :show, location))
+          |> render("show.json", location: location)
+        end
+      _else ->
+        conn
+        |> redirect(to: page_path(conn, :index))
     end
   end
 
@@ -34,18 +46,47 @@ defmodule StormchatWeb.LocationController do
     render(conn, "show.json", location: location)
   end
 
+  # creates a saved location for the verified user
   def update(conn, %{"id" => id, "location" => location_params}) do
-    location = Locations.get_location!(id)
+    token = assigns(conn, :token)
 
-    with {:ok, %Location{} = location} <- Locations.update_location(location, location_params) do
-      render(conn, "show.json", location: location)
+    case Phoenix.Token.verify(conn, "auth token", token, max_age: 86400) do
+      {:ok, user_id} ->
+        location = Locations.get_location!(id)
+
+        # make sure the given location is the verified user's
+        if location.user_id == user_id do
+          with {:ok, %Location{} = location} <- Locations.update_location(location, location_params) do
+            render(conn, "show.json", location: location)
+          end
+        else
+          |> redirect(to: page_path(conn, :index))
+        end
+      _else ->
+        conn
+        |> redirect(to: page_path(conn, :index))
     end
   end
 
+  # deletes a verified user's saved location
   def delete(conn, %{"id" => id}) do
-    location = Locations.get_location!(id)
-    with {:ok, %Location{}} <- Locations.delete_location(location) do
-      send_resp(conn, :no_content, "")
+    token = assigns(conn, :token)
+
+    case Phoenix.Token.verify(conn, "auth token", token, max_age: 86400) do
+      {:ok, user_id} ->
+        location = Locations.get_location!(id)
+
+        # make sure the given location is the verified user's
+        if location.user_id == user_id do
+          with {:ok, %Location{}} <- Locations.delete_location(location) do
+            send_resp(conn, :no_content, "")
+          end
+        else
+          |> redirect(to: page_path(conn, :index))
+        end
+      _else ->
+        conn
+        |> redirect(to: page_path(conn, :index))
     end
   end
 end
